0.1.4
dispose 기능 버그 수정
- dispose가 disposed로 잘못되어있어서 dispose를 정정함
- 기존에는 firebase 컴포넌트가 제거되지 않는 문제가 있었

+++ src/firebase.js
@@ -134,10 +134,13 @@ export default class Firebase extends RectPath(Shape) {
     promise.catch(e => console.log(e.message))
   }
 
-  disposed() {
-    firebase.auth().signOut();
-    firebase.app().delete();
-    // this._database && ..
+  dispose() {
+    if(this.app.isViewMode) {
+      firebase.auth().signOut()
+      firebase.app().delete();
+    }
+
+    super.dispose()
   }
 
   _draw(context) {
